requirements.txt
flask
twilio
google-generativeai
flask-sqlalchemy
requests
gunicorn